Change model loading test to be agnostic about tensorflow/keras version
The current test was failing due to differences in how models are
sumamrised. This new test instead checks that weights and model
structure is as expected, without worrying about the underlying object
looks like.

This probably isn't the perfect way of doing this, but I think it is
sufficient. Importantly, it should no longer break (so easily) following
updates to tensorflow/keras.